+++ packages/liveblocks-node/src/__tests__/client.test.ts
@@ -1,3 +1,8 @@
+import type {
+  CommentData,
+  CommentUserReaction,
+  ThreadData,
+} from "@liveblocks/core";
 import { http, HttpResponse } from "msw";
 import { setupServer } from "msw/node";
 
@@ -35,6 +40,45 @@ describe("client", () => {
       },
     },
   ];
+
+  const comment: CommentData = {
+    type: "comment",
+    id: "comment1",
+    roomId: "room1",
+    threadId: "thread1",
+    userId: "user1",
+    createdAt: new Date("2022-07-13T14:32:50.697Z"),
+    reactions: [
+      {
+        emoji: "🐐",
+        createdAt: new Date("2022-07-13T14:32:50.697Z"),
+        users: [{ id: "user1" }],
+      },
+    ],
+    body: {
+      version: 1,
+      content: [],
+    },
+  };
+
+  const thread: ThreadData<{
+    color: string;
+  }> = {
+    type: "thread",
+    id: "thread1",
+    roomId: "room1",
+    metadata: {
+      color: "blue",
+    },
+    createdAt: new Date("2022-07-13T14:32:50.697Z"),
+    comments: [comment],
+  };
+
+  const reaction: CommentUserReaction = {
+    emoji: "🐐",
+    createdAt: new Date("2022-07-13T14:32:50.697Z"),
+    userId: "user1",
+  };
 
   const server = setupServer(
     http.get(`${DEFAULT_BASE_URL}/v2/rooms`, () => {
@@ -157,7 +201,178 @@ describe("client", () => {
     });
   });
 
-  test("should throw an ApiError when getRoom receives an error response", async () => {
+  test("should return the edited comment when editComment receives a successful response", async () => {
+    const commentData = {
+      body: {
+        version: 1 as const,
+        content: [],
+      },
+      editedAt: new Date(),
+    };
+
+    server.use(
+      http.post(
+        `${DEFAULT_BASE_URL}/v2/rooms/:roomId/threads/:threadId/comments/:commentId`,
+        async ({ request }) => {
+          const data = await request.json();
+
+          if (JSON.stringify(data) === JSON.stringify(commentData)) {
+            return HttpResponse.json(comment);
+          }
+
+          return HttpResponse.error();
+        }
+      )
+    );
+
+    const client = new Liveblocks({ secret: "sk_xxx" });
+    const res = await client.editComment({
+      roomId: "room1",
+      threadId: "thread1",
+      commentId: "comment1",
+      data: commentData,
+    });
+    expect(res).toEqual(comment);
+  });
+
+  test("should throw a LiveblocksError when editComment receives an error response", async () => {
+    const error = {
+      error: "RESOURCE_NOT_FOUND",
+      message: "Comment not found",
+    };
+
+    server.use(
+      http.post(
+        `${DEFAULT_BASE_URL}/v2/rooms/:roomId/threads/:threadId/comments/:commentId`,
+        () => {
+          return HttpResponse.json(error, { status: 404 });
+        }
+      )
+    );
+
+    const client = new Liveblocks({ secret: "sk_xxx" });
+
+    // This should throw a LiveblocksError
+    try {
+      // Attempt to get, which should fail and throw an error.
+      await client.editComment({
+        roomId: "room1",
+        threadId: "thread1",
+        commentId: "comment1",
+        data: {
+          body: {
+            version: 1 as const,
+            content: [],
+          },
+        },
+      });
+      // If it doesn't throw, fail the test.
+      expect(true).toBe(false);
+    } catch (err) {
+      expect(err instanceof LiveblocksError).toBe(true);
+      if (err instanceof LiveblocksError) {
+        expect(err.status).toBe(404);
+        expect(err.message).toBe(JSON.stringify(error));
+        expect(err.name).toBe("LiveblocksError");
+      }
+    }
+  });
+
+  test("should return the created thread when createThread receives a successful response", async () => {
+    const threadData = {
+      comment: {
+        userId: "user1",
+        createdAt: new Date(),
+        body: {
+          version: 1 as const,
+          content: [],
+        },
+      },
+      metadata: {
+        color: "blue",
+      },
+    };
+
+    server.use(
+      http.post(
+        `${DEFAULT_BASE_URL}/v2/rooms/:roomId/threads`,
+        async ({ request }) => {
+          const data = await request.json();
+
+          if (JSON.stringify(threadData) === JSON.stringify(data)) {
+            return HttpResponse.json(thread);
+          }
+
+          return HttpResponse.error();
+        }
+      )
+    );
+
+    const client = new Liveblocks({ secret: "sk_xxx" });
+    const res = await client.createThread({
+      roomId: "room1",
+      data: threadData,
+    });
+
+    expect(res).toEqual(thread);
+  });
+
+  test("should throw a LiveblocksError when createThread receives an error response", async () => {
+    const threadData = {
+      comment: {
+        userId: "user1",
+        createdAt: new Date(),
+        body: {
+          version: 1 as const,
+          content: [],
+        },
+      },
+      metadata: {
+        color: "blue",
+      },
+    };
+
+    const error = {
+      error: "RESOURCE_ALREADY_EXISTES",
+      message: "Thread already exists",
+    };
+
+    server.use(
+      http.post(
+        `${DEFAULT_BASE_URL}/v2/rooms/:roomId/threads`,
+        async ({ request }) => {
+          const data = await request.json();
+          if (JSON.stringify(threadData) === JSON.stringify(data)) {
+            return HttpResponse.json(error, { status: 409 });
+          }
+
+          return HttpResponse.error();
+        }
+      )
+    );
+
+    const client = new Liveblocks({ secret: "sk_xxx" });
+
+    // This should throw a LiveblocksError
+    try {
+      // Attempt to create a thread, which should fail and throw an error.
+      await client.createThread({
+        roomId: "room1",
+        data: threadData,
+      });
+      // If it doesn't throw, fail the test.
+      expect(true).toBe(false);
+    } catch (err) {
+      expect(err instanceof LiveblocksError).toBe(true);
+      if (err instanceof LiveblocksError) {
+        expect(err.status).toBe(409);
+        expect(err.message).toBe(JSON.stringify(error));
+        expect(err.name).toBe("LiveblocksError");
+      }
+    }
+  });
+
+  test("should throw a LiveblocksError when getRoom receives an error response", async () => {
     const error = {
       error: "ROOM_NOT_FOUND",
       message: "Room not found",
@@ -174,7 +389,7 @@ describe("client", () => {
 
     const client = new Liveblocks({ secret: "sk_xxx" });
 
-    // This should throw an HttpError
+    // This should throw an LiveblocksError
     try {
       // Attempt to get, which should fail and throw an error.
       await client.getRoom("123");
@@ -188,6 +403,89 @@ describe("client", () => {
         expect(err.name).toBe("LiveblocksError");
       }
     }
+  });
+
+  test("should return a list of threads when getThreads receives a successful response", async () => {
+    server.use(
+      http.get(`${DEFAULT_BASE_URL}/v2/rooms/:roomId/threads`, () => {
+        return HttpResponse.json(
+          {
+            data: [thread],
+          },
+          { status: 200 }
+        );
+      })
+    );
+
+    const client = new Liveblocks({ secret: "sk_xxx" });
+
+    await expect(
+      client.getThreads({
+        roomId: "room1",
+      })
+    ).resolves.toEqual({
+      data: [thread],
+    });
+  });
+
+  test("should return the specified thread when getThread receives a successful response", async () => {
+    server.use(
+      http.get(`${DEFAULT_BASE_URL}/v2/rooms/:roomId/threads/:threadId`, () => {
+        return HttpResponse.json(thread, { status: 200 });
+      })
+    );
+
+    const client = new Liveblocks({ secret: "sk_xxx" });
+
+    await expect(
+      client.getThread({
+        roomId: "room1",
+        threadId: "thread1",
+      })
+    ).resolves.toEqual(thread);
+  });
+
+  test("should return the specified comment when getComment receives a successful response", async () => {
+    server.use(
+      http.get(
+        `${DEFAULT_BASE_URL}/v2/rooms/:roomId/threads/:threadId/comments/:commentId`,
+        () => {
+          return HttpResponse.json(comment, { status: 200 });
+        }
+      )
+    );
+
+    const client = new Liveblocks({ secret: "sk_xxx" });
+
+    await expect(
+      client.getComment({
+        roomId: "room1",
+        threadId: "thread1",
+        commentId: "comment1",
+      })
+    ).resolves.toEqual(comment);
+  });
+
+  test("should add a reaction to a comment when addReaction receives a successful response", async () => {
+    server.use(
+      http.post(
+        `${DEFAULT_BASE_URL}/v2/rooms/:roomId/threads/:threadId/comments/:commentId/add-reaction`,
+        () => {
+          return HttpResponse.json(reaction, { status: 200 });
+        }
+      )
+    );
+
+    const client = new Liveblocks({ secret: "sk_xxx" });
+
+    await expect(
+      client.addCommentReaction({
+        roomId: "room1",
+        threadId: "thread1",
+        commentId: "comment1",
+        data: reaction,
+      })
+    ).resolves.toEqual(reaction);
   });
 
   test("should throw an error when getRoom fails due to network error", async () => {