+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### `@liveblocks/node`
 
+- Implement Comments Write REST APIs as fully typed methods. (includes
+  `createThread`, `editThreadMetadata`, `createComment`, `editComment`,
+  `deleteComment`, `addCommentReaction`, and `removeCommentReaction` methods)
 - Fixes the return type of `getActiveUsers` to match the data returned from the
   endpoint.